Map coming together.. Sorta
I put a little village in the game and i put a few shrubs in there. I
found a lot more grassy sprites and firey sprites and a few lightning
sprites.. stilll looking for dark themed sprites. Anyway, not even close
to done with the map. Just didnt have much time because I started this
tonight